v2.2: завершение всех задач (TASK-028..TASK-038) — ARP-резолвер, плагины, GUI, права, документация

+++ internal/gui/accent_colors.go
@@ -91,7 +91,7 @@ func (a *App) saveAccentColor(name string, c color.RGBA) {
 		return
 	}
 	p := a.myApp.Preferences()
-	val := c.R<<24 | c.G<<16 | c.B<<8 | c.A
+	val := uint32(c.R)<<24 | uint32(c.G)<<16 | uint32(c.B)<<8 | uint32(c.A)
 	p.SetString(accentKey(name), strconv.FormatUint(uint64(val), 16))
 }
